Allows external secrets for Postgres and Redis (#1)
TL;DR
-----

Allows user to specify existing secrets for Redis and Postgres

Details
-------

Providse new values to support using existing secrets for the
Postgres and Redis passwords. I did this in preparation for 
insalling the chart with ArgoCD using sealed secrets. It will
also be useful moving toward using external databases which is
the next change.

Also makes some improvements to the build and CI processes to
polish them up a bit.